Added: new ``clientScripts`` option to disable all JavaScript code 🙃
This allows you to turn Phenomic into a standard static website
generator

Closes #874
