+++ flow/interfaces/webpack.js
@@ -1,2 +1,4 @@
 // @todo improve this
 export type WebpackConfig = Object
+
+export type WebpackAssetsFiles = { [key: string]: Array<string> }
