+++ src/static/__tests__/fixtures/collection.js
