+++ src/static/__tests__/fixtures/routesWithCustomScript.js
