+++ src/static/__tests__/fixtures/store.js
